+++ run_tests.sh
@@ -26,12 +26,14 @@ function usage {
   echo "  -n, --no-recreate-db     Don't recreate the test database."
   echo "  -x, --stop               Stop running tests after the first error or failure."
   echo "  -f, --force              Force a clean re-build of the virtual environment. Useful when dependencies have been added."
+  echo "  -u, --update             Update the virtual environment with any newer package versions"
   echo "  -p, --pep8               Just run pep8"
   echo "  -P, --no-pep8            Don't run pep8"
   echo "  -c, --coverage           Generate coverage report"
   echo "  -h, --help               Print this usage message"
   echo "  -xintegration            Ignore all keystoneclient test cases (integration tests)"
   echo "  --hide-elapsed           Don't print the elapsed time for each test along with slow test list"
+  echo "  --standard-threads       Don't do the eventlet threading monkeypatch."
   echo ""
   echo "Note: with no options specified, the script will try to run the tests in a virtual environment,"
   echo "      If no virtualenv is found, the script will ask if you would like to create one.  If you "
@@ -47,10 +49,15 @@ function process_option {
     -r|--recreate-db) recreate_db=1;;
     -n|--no-recreate-db) recreate_db=0;;
     -f|--force) force=1;;
+    -u|--update) update=1;;
     -p|--pep8) just_pep8=1;;
+    -8|--8) short_pep8=1;;
     -P|--no-pep8) no_pep8=1;;
     -c|--coverage) coverage=1;;
-	-xintegration) nokeystoneclient=1;;
+    -xintegration) nokeystoneclient=1;;
+    --standard-threads)
+        export STANDARD_THREADS=1
+        ;;
     -*) noseopts="$noseopts $1";;
     *) noseargs="$noseargs $1"
   esac
@@ -65,10 +72,12 @@ noseargs=
 noseopts=
 wrapper=""
 just_pep8=0
+short_pep8=0
 no_pep8=0
 coverage=0
 nokeystoneclient=0
 recreate_db=1
+update=0
 
 for arg in "$@"; do
   process_option $arg
@@ -80,13 +89,13 @@ if [ $coverage -eq 1 ]; then
 fi
 
 if [ $nokeystoneclient -eq 1 ]; then
-	# disable the integration tests
+    # disable the integration tests
     noseopts="$noseopts -I test_keystoneclient*"
 fi
 
 function run_tests {
   # Just run the test suites in current environment
-  ${wrapper} $NOSETESTS 2> run_tests.log
+  ${wrapper} $NOSETESTS
   # If we get some short import error right away, print the error log directly
   RESULT=$?
   if [ "$RESULT" -ne "0" ];
@@ -101,20 +110,27 @@ function run_tests {
 }
 
 function run_pep8 {
+  FLAGS=--show-pep8
+  echo $#
+  if [ $# -gt 0 ] && [ 'short' == ''$1 ]
+  then
+      FLAGS=''
+  fi
+
+
   echo "Running pep8 ..."
   # Opt-out files from pep8
-  ignore_scripts="*.sh:"
-  ignore_files="*eventlet-patch:*pip-requires"
-  ignore_dirs="*ajaxterm*"
-  GLOBIGNORE="$ignore_scripts:$ignore_files:$ignore_dirs"
-  srcfiles=`find bin -type f ! -name .*.swp`
-  srcfiles+=" keystone"
+  ignore_scripts="*.pyc,*.pyo,*.sh,*.swp,*.rst"
+  ignore_files="*pip-requires"
+  ignore_dirs=".venv,.tox,dist,doc,openstack,vendor,*egg"
+  ignore="$ignore_scripts,$ignore_files,$ignore_dirs"
+  srcfiles="."
   # Just run PEP8 in current environment
-  ${wrapper} pep8 --repeat --show-pep8 --show-source \
-    --exclude=vcsversion.py ${srcfiles} | tee pep8.txt
+  ${wrapper} pep8 --repeat $FLAGS --show-source \
+    --exclude=${ignore} ${srcfiles} | tee pep8.txt
 }
 
-NOSETESTS="python run_tests.py $noseopts $noseargs"
+NOSETESTS="nosetests $noseopts $noseargs"
 
 if [ $never_venv -eq 0 ]
 then
@@ -122,6 +138,10 @@ then
   if [ $force -eq 1 ]; then
     echo "Cleaning virtualenv..."
     rm -rf ${venv}
+  fi
+  if [ $update -eq 1 ]; then
+    echo "Updating virtualenv..."
+    python tools/install_venv.py
   fi
   if [ -e ${venv} ]; then
     wrapper="${with_venv}"
@@ -151,6 +171,12 @@ if [ $just_pep8 -eq 1 ]; then
     run_pep8
     exit
 fi
+
+if [ $short_pep8 -eq 1 ]; then
+     run_pep8 short
+     exit
+fi
+
 
 if [ $recreate_db -eq 1 ]; then
     rm -f tests.sqlite
